// Angular/Generic Modules/extensible-data-table/Enums/filter-types.enum.ts
export enum FilterTypes {
    Contains = 'Contains',
    EndsWith = 'Ends With',
    Equals = 'Equals',
    GreaterThan = 'Greater Than',
    GraterThanOrEqual = 'Greater Than or Equal',
    LessThan = 'Less Than',
    LessThanOrEqualTo = 'Less Than or Equal To',
    NotEqual = 'Not Equal',
    StartsWith = 'Starts With',
}